WebIDL/tests/submissions/heycam/constant-001.html

<!DOCTYPE html>
<title>Constants on interfaces</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.5.5 Constants
//
// Constants are exposed as non-writable, enumerable, non-configurable
// properties on interface objects and interface prototype objects.

// feature: HTMLMediaElement.NETWORK_IDLE (constant)

test(function() {
  assert_property(HTMLMediaElement, "NETWORK_IDLE", { writable: false, enumerable: true, configurable: false, value: 1 });
}, "HTMLMediaElement.NETWORK_IDLE has correct attributes and value");

test(function() {
  assert_property(HTMLMediaElement.prototype, "NETWORK_IDLE", { writable: false, enumerable: true, configurable: false, value: 1 });
}, "HTMLMediaElement.prototype.NETWORK_IDLE has correct attributes and value");

// feature: SVGLength.SVG_LENGTHTYPE_NUMBER (constant)

test(function() {
  assert_property(SVGLength, "SVG_LENGTHTYPE_NUMBER", { writable: false, enumerable: true, configurable: false, value: 1 });
}, "SVGLength.SVG_LENGTHTYPE_NUMBER has correct attributes and value");

test(function() {
  assert_property(SVGLength.prototype, "SVG_LENGTHTYPE_NUMBER", { writable: false, enumerable: true, configurable: false, value: 1 });
}, "SVGLength.prototype.SVG_LENGTHTYPE_NUMBER has correct attributes and value");
</script>
